Refactor: unifying Login and RegisterPage Styles
Adding responsiveness from 900 to 425px
